fixed catalog serialization, bumped version of library, fixed mql and cypher

+++ plugins/sql-language/src/main/java/org/polypheny/db/sql/language/ddl/altertable/SqlAlterTableDropColumn.java
@@ -20,6 +20,7 @@ package org.polypheny.db.sql.language.ddl.altertable;
 import java.util.List;
 import java.util.Objects;
 import org.polypheny.db.catalog.entity.logical.LogicalTable;
+import org.polypheny.db.catalog.exceptions.GenericRuntimeException;
 import org.polypheny.db.catalog.logistic.EntityType;
 import org.polypheny.db.ddl.DdlManager;
 import org.polypheny.db.languages.ParserPos;
@@ -77,12 +78,16 @@ public class SqlAlterTableDropColumn extends SqlAlterTable {
     public void execute( Context context, Statement statement, QueryParameters parameters ) {
         LogicalTable catalogTable = getEntityFromCatalog( context, table );
 
+        if ( catalogTable == null ) {
+            throw new GenericRuntimeException( "Entity with name '%s' was not found.", String.join( ".", table.names ) );
+        }
+
         if ( catalogTable.entityType != EntityType.ENTITY && catalogTable.entityType != EntityType.SOURCE ) {
             throw new RuntimeException( "Not possible to use ALTER TABLE because " + catalogTable.name + " is not a table." );
         }
 
         if ( column.names.size() != 1 ) {
-            throw new RuntimeException( "No FQDN allowed here: " + column.toString() );
+            throw new RuntimeException( "No FQDN allowed here: " + column );
         }
 
         DdlManager.getInstance().dropColumn( catalogTable, column.getSimple(), statement );
